requirements.txt
Flask==2.3.4
Flask-MySQLdb==1.0.1
pandas==2.1.0
Pillow==10.0.0
requests==2.30.0
